feat: allow disabling the logger via a feature
